# (#6031) - faster IDB changes() with batched cursor