@daleharvey 2016-12-21
That is a huge speedup, awesome job 👍